+++ src/Plugin.cs
@@ -9,7 +9,7 @@ using FFUIOverhaul.Utils;
 using FFUIOverhaul.TechTree;
 using FFUIOverhaul.Settings;
 
-[assembly: MelonInfo(typeof(FFUIOverhaul.FFUIOverhaulMod), "Keep Clarity", "1.2.2", "sagedragoon79")]
+[assembly: MelonInfo(typeof(FFUIOverhaul.FFUIOverhaulMod), "Keep Clarity", "1.2.3", "sagedragoon79")]
 [assembly: MelonGame("Crate Entertainment", "Farthest Frontier")]
 
 namespace FFUIOverhaul
@@ -420,7 +420,6 @@ namespace FFUIOverhaul
             HandlePauseOnLoad(gm);
             Patches.DismissibleResourceAlerts.Tick();
             UI.CompanyOverlayManager.Tick();
-            Patches.PredatorBlurbBackground.Tick(Time.unscaledDeltaTime);
         }
 
         private void HandlePauseOnLoad(GameManager gm)